Refactor wallet specs to improve separation between construction and test logic
